requirements.txt
configobj
NamedAtomicLock>=1.1.0
func_timeout
python-subprocess2
